worker/requirements.txt
# Worker Requirements
celery[redis]==5.3.6
redis==5.0.1

# Image Processing
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80

# File Processing
potrace==1.8

# Database
supabase==2.3.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
